refactor: add current provider impl
