+++ src/main/java/br/com/thiagoRDS/api_authors/modules/authors/usecases/ForgotMailPassword.java
@@ -10,7 +10,7 @@ import br.com.thiagoRDS.api_authors.modules.authors.entities.RecoveryToken;
 import br.com.thiagoRDS.api_authors.modules.authors.exceptions.AuthorNotFoundException;
 import br.com.thiagoRDS.api_authors.modules.authors.repositories.AuthorsRepository;
 import br.com.thiagoRDS.api_authors.modules.authors.repositories.RecoveryTokensRepository;
-import br.com.thiagoRDS.api_authors.providers.CurrentContextProvider;
+import br.com.thiagoRDS.api_authors.providers.CurrentContextProvider.CurrentContextProvider;
 import br.com.thiagoRDS.api_authors.providers.MailProvider.MailProvider;
 import br.com.thiagoRDS.api_authors.providers.MailProvider.dtos.SendMailDTO;
